refactor: consolidate Interface and Theme components into UserInterface, enhancing settings management
